+++ linkify-widget.php
@@ -120,20 +120,17 @@ abstract class c2c_LinkifyWidget extends WP_Widget {
 			$$key = apply_filters( $this->widget_id . '_widget_' . $key, $instance[ $key ] );
 		}
 
-		// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-		echo $before_widget;
+		echo wp_kses_post( $before_widget );
 
 		if ( $title ) {
-			// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-			echo $before_title . $title . $after_title;
+			echo wp_kses_post( $before_title . $title . $after_title );
 		}
 
 		// Widget content
 		$args = compact( array_keys( $this->config ) );
 		$this->widget_content( $args, $instance );
 
-		// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-		echo $after_widget;
+		echo wp_kses_post( $after_widget );
 	}
 
 	public function update( $new_instance, $old_instance ) {
@@ -301,8 +298,9 @@ abstract class c2c_LinkifyWidget extends WP_Widget {
 				);
 			}
 			if ( $this->config[ $opt ]['help'] ) {
-				// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-				echo "<div style='color:#888; font-size:x-small;'>({$this->config[ $opt ]['help']})</div>";
+				echo '<div style="color:#888; font-size:x-small;">(';
+				echo wp_kses_post( $this->config[ $opt ]['help'] );
+				echo ')</div>';
 			}
 			echo "</p>\n";
 		}